requirements.txt
pygame
psycopg2-binary
flask
flask-cors
